memovision_backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
numpy==1.26.4
pandas==2.2.3
chromadb
requests
Pillow==10.3
open-clip-torch
pybase64
python-dotenv==1.0.0
fastapi-cors
groq
